Refactor machines tests
